Karnataka forest dataset

Dataset from 96 forest plots (1 ha) established in the central Western Ghats of India by Ramesh et al. (2010).

Usage

data("KarnatakaForest")

Format

A data frame with 65889 observations on the following 8 variables :

plotId: Names of the plots

treeId: Tree Id, contains a letter (A, B, C...) when an individual has multiple stems

family: Family

genus: Genus

species: Species

D: Diameter (cm)

lat: Latitude

long: Longitude

References

Ramesh, B. R. et al. (2010). Forest stand structure and composition in 96 sites along environmental gradients in the central Western Ghats of India Ecological Archives E091-216. Ecology, 91(10), 3118-3118.

Examples

data(KarnatakaForest)
str(KarnatakaForest)
#> 'data.frame':	65889 obs. of  8 variables:
#>  $ plotId : chr  "BSP1" "BSP1" "BSP1" "BSP1" ...
#>  $ treeId : chr  "BSP1_1" "BSP1_2" "BSP1_3" "BSP1_4" ...
#>  $ family : chr  "Fabaceae" "Fabaceae" "Fabaceae" "Fabaceae" ...
#>  $ genus  : chr  "Acacia" "Acacia" "Acacia" "Acacia" ...
#>  $ species: chr  "sinuata" "sinuata" "sinuata" "sinuata" ...
#>  $ D      : num  3.82 4.14 4.14 4.46 4.46 ...
#>  $ lat    : num  14.4 14.4 14.4 14.4 14.4 ...
#>  $ long   : num  74.9 74.9 74.9 74.9 74.9 ...