Function that return a possibly cached file, transparently downloading it if missing

Function that return a possibly cached file, transparently downloading it if missing

Usage

cacheManager(nameFile)

Arguments

nameFile

character. file to resolve cached path.

Value

file path of the resolved cached file.

Localisation

Cache path discovery protocol

BIOMASS.cache option set to an existing folder

existing user data folder rappdirs::user_data_dir()

On Linux : ~/.local/share/R/BIOMASS

On Mac OS X : ~/Library/Application Support/R/BIOMASS

On Windows 7 up to 10 : C:\\Users\\<username>\\AppData\\Local\\R\\BIOMASS

On Windows XP : C:\\Documents and Settings\\<username>\\Data\\R\\BIOMASS

fallback to R session tempdir

Author

Guillaume CORNU