Divides one ore more plots into subplots

This function divides a plot (or several plots) into subplots in the relative coordinates system, and returns the coordinates of subplot corners.

Usage

divide_plot(
  rel_coord,
  proj_coord = NULL,
  grid_size,
  tree_data = NULL,
  tree_coords = NULL,
  corner_plot_ID = NULL,
  tree_plot_ID = NULL,
  grid_tol = 0.1,
  centred_grid = F
)

Arguments

rel_coord

a matrix or data frame containing the relative (local) coordinates of plot corners, with X and Y corresponding to the first and second column respectively

proj_coord

a matrix or data frame containing the projected coordinates of plot corners, with X and Y corresponding to the first and second column respectively, and with the same row order than rel_coord

grid_size

a vector indicating the dimensions of grid cells (resp. X and Y dimensions). If only one value is given, grid cells will be considered as squares.

tree_data

a data frame containing tree relative coordinates and other tree metrics (one row per tree).

tree_coords

a character vector of size 2 containing the column names of the relative coordinates of the trees.

corner_plot_ID

if dealing with multiple plots : a vector indicating plot IDs for corners.

tree_plot_ID

if dealing with multiple plots : a vector indicating tree plot IDs.

grid_tol

a numeric between (0;1) corresponding to the percentage of the plot area allowed to be excluded from the plot division (when grid_size doesn't match exactly plot dimensions).

centred_grid

when grid_size doesn't match exactly plot dimensions, a logical indicating if the subplot grid should be centred on the plot.

Value

If tree_data isn't supplied, returns a data-frame containing as many rows as there are corners corresponding to the subplots, and the following columns :

corner_plot_ID: If dealing with multiple plots : the plot code

subplot_id: The automatically generated subplot code, using the following rule : subplot_X_Y

x_rel and y_rel : the relative X-axis and Y-axis coordinates of subplots corners.

x_proj and y_proj : if proj_coord is supplied, the projected X-axis and Y-axis coordinates of subplots corners

If tree_data is supplied, returns a list containing :

the previous data-frame

the tree_data data-frame with the subplot_id of each tree in the last column

Details

If corner coordinates in the projected coordinate system are supplied (proj_coord), projected coordinates of subplot corners are calculated by a bilinear interpolation in relation with relative coordinates of plot corners. Be aware that this bilinear interpolation only works if the plot in the relative coordinates system is rectangular (ie, has 4 right angles).

Author

Arthur PERE, Arthur BAILLY

Examples

# Rectangular plot and grid cells
rel_coord <- data.frame(x_rel = c(0, 200, 0, 200), y_rel = c(0, 0, 100, 100))
subplots <- divide_plot(rel_coord = rel_coord, grid_size = c(100,50))

# Squared plot and projected coordinates associated
rel_coord <- data.frame(x_rel = c(0, 200, 0, 200), y_rel = c(0, 0, 200, 200))
proj_coord <- data.frame(x_proj = c(210, 383, 110, 283), y_proj = c(210, 310, 383, 483))
subplots <- divide_plot(rel_coord, proj_coord = proj_coord, grid_size = 100)

# Assigning trees to subplots
tree_data <- data.frame(x_tree = runif(50,0,200), y_tree = runif(50,0,200))
subplots <- divide_plot(rel_coord, proj_coord, 100,
                        tree_data = tree_data, tree_coords = c("x_tree","y_tree"))
subplots$sub_corner_coord
#>     subplot_id x_rel y_rel x_proj y_proj
#> 1  subplot_0_0     0     0  210.0  210.0
#> 2  subplot_0_0   100     0  296.5  260.0
#> 3  subplot_0_0   100   100  246.5  346.5
#> 4  subplot_0_0     0   100  160.0  296.5
#> 5  subplot_1_0   100     0  296.5  260.0
#> 6  subplot_1_0   200     0  383.0  310.0
#> 7  subplot_1_0   200   100  333.0  396.5
#> 8  subplot_1_0   100   100  246.5  346.5
#> 9  subplot_0_1     0   100  160.0  296.5
#> 10 subplot_0_1   100   100  246.5  346.5
#> 11 subplot_0_1   100   200  196.5  433.0
#> 12 subplot_0_1     0   200  110.0  383.0
#> 13 subplot_1_1   100   100  246.5  346.5
#> 14 subplot_1_1   200   100  333.0  396.5
#> 15 subplot_1_1   200   200  283.0  483.0
#> 16 subplot_1_1   100   200  196.5  433.0
subplots$tree_data
#>        x_tree     y_tree  subplot_id
#> 1   51.958106  74.680080 subplot_0_0
#> 2   14.223925 145.035017 subplot_0_1
#> 3    3.843068 160.729281 subplot_0_1
#> 4   46.451866 148.842696 subplot_0_1
#> 5  164.357939 122.537341 subplot_1_1
#> 6   83.782042  73.838098 subplot_0_0
#> 7   43.705988 110.826181 subplot_0_1
#> 8  116.631934  67.935801 subplot_1_0
#> 9    2.773768 178.235419 subplot_0_1
#> 10 163.972499  45.357006 subplot_1_0
#> 11 123.461906  49.478801 subplot_1_0
#> 12 163.245088  22.208424 subplot_1_0
#> 13 190.140255 105.218617 subplot_1_1
#> 14 132.747603  66.321216 subplot_1_0
#> 15 127.047749  15.298431 subplot_1_0
#> 16   1.593114 101.468876 subplot_0_1
#> 17 137.033055 123.120472 subplot_1_1
#> 18   9.646664 114.523066 subplot_0_1
#> 19 112.610381  42.983369 subplot_1_0
#> 20  89.237819 171.052247 subplot_0_1
#> 21  95.743520   7.940341 subplot_0_0
#> 22 183.652175  84.009961 subplot_1_0
#> 23 164.058545  48.652797 subplot_1_0
#> 24  51.210288 118.042125 subplot_0_1
#> 25 168.134278 173.901249 subplot_1_1
#> 26  44.234386 169.742622 subplot_0_1
#> 27  23.736516 114.831022 subplot_0_1
#> 28  69.994976  36.474899 subplot_0_0
#> 29 148.528853 186.509067 subplot_1_1
#> 30 133.215365  18.512889 subplot_1_0
#> 31 128.947254 139.381012 subplot_1_1
#> 32 157.410616 101.205512 subplot_1_1
#> 33  40.794233 121.014920 subplot_0_1
#> 34  44.637811 154.445944 subplot_0_1
#> 35   6.310994 167.263967 subplot_0_1
#> 36 164.111599  94.343070 subplot_1_0
#> 37  54.215509 194.434294 subplot_0_1
#> 38 189.245818 127.284878 subplot_1_1
#> 39  65.223722 165.754952 subplot_0_1
#> 40  64.314446 101.065343 subplot_0_1
#> 41 143.684919   8.797394 subplot_1_0
#> 42 116.168296 184.187304 subplot_1_1
#> 43  28.926403 135.003247 subplot_0_1
#> 44 193.006966  23.452007 subplot_1_0
#> 45   7.390662  79.256956 subplot_0_0
#> 46  14.931466 117.729061 subplot_0_1
#> 47  93.548940  26.158600 subplot_0_0
#> 48 118.595818  32.430279 subplot_1_0
#> 49  11.012697 175.333927 subplot_0_1
#> 50  64.757544  62.799655 subplot_0_0

# When grid dimensions don't fit perfectly plot dimensions
# \donttest{
 subplots <- divide_plot(rel_coord, grid_size = c(45,50))
#> Warning: The x-dimension of the plot is not a multiple of the x-dimension of the grid size
 subplots <- divide_plot(rel_coord, grid_size = c(41,41) , grid_tol = 0.4, centred_grid = TRUE)
#> Warning: The x-dimension of the plot is not a multiple of the x-dimension of the grid size
#> Warning: The y-dimension of the plot is not a multiple of the y-dimension of the grid size
# }

# Dealing with multiple plots
rel_coord <- rbind(rel_coord, rel_coord)
proj_coord <- rbind(proj_coord, proj_coord + 200)
tree_data <- rbind(tree_data, data.frame(x_tree = runif(50,0,200), y_tree = runif(50,0,200)))
corner_plot_ID <- rep(c("plot1","plot2"), e=4)
tree_plot_ID <- rep(c("plot1","plot2"), e=50)
subplots <- divide_plot(rel_coord, proj_coord, 100,
                        tree_data, c("x_tree","y_tree"),
                        corner_plot_ID, tree_plot_ID)