Summarise and display tree information by subplot

After applying the divide_plot() function, this function summarises with any defined function the desired tree metric by sub-plot and displays the plot representation.

Usage

subplot_summary(subplots, value = NULL, draw_plot = TRUE, fun = sum, ...)

Arguments

subplots

output of the divide_plot() function

value

a character indicating the column in subplots$tree_data to display

draw_plot

a logical indicating whether the plot design should be displayed and returned

fun

the function to be applied

...

optional arguments to fun

Value

a list containg the following elements :

tree_summary : a summary of the metric per subplot

polygon : an sf object : simple feature collection of the subplot's polygon

plot_design : (if draw_plot=T) a ggplot object (or a list of ggplot objects) that can easily be modified

Author

Arthur Bailly

Examples

rel_coord <- data.frame(x_rel = c(0, 200, 0, 200), y_rel = c(0, 0, 200, 200))
proj_coord <- data.frame(x_proj = c(210, 383, 110, 283), y_proj = c(210, 310, 383, 483))
tree_data <- data.frame(x_tree = runif(50,0,200), y_tree = runif(50,0,200), metric = rnorm(50,10,5))
subplots <- divide_plot(rel_coord, proj_coord = proj_coord, grid_size = 50,
                        tree_data = tree_data, tree_coords = c("x_tree","y_tree"))
# Sum summary (by default)
subplot_summary(subplots , value = "metric", draw_plot = FALSE)
#> $tree_summary
#>     subplot_id metric_summary metric_summary_per_ha
#> 1  subplot_0_0       24.49649              98.16018
#> 2  subplot_0_1       30.05625             120.43879
#> 3  subplot_0_2       90.13604             361.18527
#> 4  subplot_0_3       24.12797              96.68348
#> 5  subplot_1_0       18.03139              72.25381
#> 6  subplot_1_1       12.30719              49.31628
#> 7  subplot_1_2       15.96526              63.97458
#> 8  subplot_1_3       15.45064              61.91244
#> 9  subplot_2_0       33.56860             134.51315
#> 10 subplot_2_1       59.52565             238.52599
#> 11 subplot_2_2       32.68268             130.96318
#> 12 subplot_2_3       20.04093              80.30626
#> 13 subplot_3_0       38.99547             156.25925
#> 14 subplot_3_1       33.77228             135.32933
#> 15 subplot_3_2       35.30392             141.46676
#> 16 subplot_3_3       37.70724             151.09714
#> 
#> $polygon
#> Simple feature collection with 16 features and 3 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 110 ymin: 210 xmax: 383 ymax: 483
#> CRS:           NA
#> First 10 features:
#>              subplot_id metric_summary metric_summary_per_ha
#> subplot_0_0 subplot_0_0       24.49649              98.16018
#> subplot_0_1 subplot_0_1       30.05625             120.43879
#> subplot_0_2 subplot_0_2       90.13604             361.18527
#> subplot_0_3 subplot_0_3       24.12797              96.68348
#> subplot_1_0 subplot_1_0       18.03139              72.25381
#> subplot_1_1 subplot_1_1       12.30719              49.31628
#> subplot_1_2 subplot_1_2       15.96526              63.97458
#> subplot_1_3 subplot_1_3       15.45064              61.91244
#> subplot_2_0 subplot_2_0       33.56860             134.51315
#> subplot_2_1 subplot_2_1       59.52565             238.52599
#>                         sf_subplot_polygon
#> subplot_0_0 POLYGON ((210 210, 253.25 2...
#> subplot_0_1 POLYGON ((185 253.25, 228.2...
#> subplot_0_2 POLYGON ((160 296.5, 203.25...
#> subplot_0_3 POLYGON ((135 339.75, 178.2...
#> subplot_1_0 POLYGON ((253.25 235, 296.5...
#> subplot_1_1 POLYGON ((228.25 278.25, 27...
#> subplot_1_2 POLYGON ((203.25 321.5, 246...
#> subplot_1_3 POLYGON ((178.25 364.75, 22...
#> subplot_2_0 POLYGON ((296.5 260, 339.75...
#> subplot_2_1 POLYGON ((271.5 303.25, 314...
#> 
# 9th quantile summary (for example)
subplot_summary(subplots , value = "metric", draw_plot = FALSE,
                fun = quantile, probs=0.9)
#> $tree_summary
#>     subplot_id metric_summary metric_summary_per_ha
#> 1  subplot_0_0      10.440282              41.83539
#> 2  subplot_0_1       8.019066              32.13330
#> 3  subplot_0_2      18.038259              72.28134
#> 4  subplot_0_3      14.515884              58.16678
#> 5  subplot_1_0      12.205742              48.90978
#> 6  subplot_1_1      12.307186              49.31628
#> 7  subplot_1_2       8.905429              35.68506
#> 8  subplot_1_3      10.620562              42.55779
#> 9  subplot_2_0      11.120390              44.56066
#> 10 subplot_2_1      17.808120              71.35914
#> 11 subplot_2_2      13.488605              54.05036
#> 12 subplot_2_3      13.808467              55.33208
#> 13 subplot_3_0      17.641247              70.69047
#> 14 subplot_3_1      13.827582              55.40868
#> 15 subplot_3_2      15.201382              60.91365
#> 16 subplot_3_3      16.271080              65.20005
#> 
#> $polygon
#> Simple feature collection with 16 features and 3 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 110 ymin: 210 xmax: 383 ymax: 483
#> CRS:           NA
#> First 10 features:
#>              subplot_id metric_summary metric_summary_per_ha
#> subplot_0_0 subplot_0_0      10.440282              41.83539
#> subplot_0_1 subplot_0_1       8.019066              32.13330
#> subplot_0_2 subplot_0_2      18.038259              72.28134
#> subplot_0_3 subplot_0_3      14.515884              58.16678
#> subplot_1_0 subplot_1_0      12.205742              48.90978
#> subplot_1_1 subplot_1_1      12.307186              49.31628
#> subplot_1_2 subplot_1_2       8.905429              35.68506
#> subplot_1_3 subplot_1_3      10.620562              42.55779
#> subplot_2_0 subplot_2_0      11.120390              44.56066
#> subplot_2_1 subplot_2_1      17.808120              71.35914
#>                         sf_subplot_polygon
#> subplot_0_0 POLYGON ((210 210, 253.25 2...
#> subplot_0_1 POLYGON ((185 253.25, 228.2...
#> subplot_0_2 POLYGON ((160 296.5, 203.25...
#> subplot_0_3 POLYGON ((135 339.75, 178.2...
#> subplot_1_0 POLYGON ((253.25 235, 296.5...
#> subplot_1_1 POLYGON ((228.25 278.25, 27...
#> subplot_1_2 POLYGON ((203.25 321.5, 246...
#> subplot_1_3 POLYGON ((178.25 364.75, 22...
#> subplot_2_0 POLYGON ((296.5 260, 339.75...
#> subplot_2_1 POLYGON ((271.5 303.25, 314...
#>