Attribute trees to GPS coordinates

Attribute trees to GPS coordinates

Usage

attributeTreeCoord(xy, plot, dim, coordAbs)

Arguments

xy

The relative coordinates of the trees within each plot

plot

The label of the plot (same length as the number of rows of xy or length of 1)

dim

The dimension of the plot (either one value if the plot is a square or a vector if a rectangle)

coordAbs

The result of the function cutPlot() or numberCorner()

Value

A data frame with two columns: - Xproj: The X coordinates in the absolute coordinate system - Yproj: The Y coordinates in the absolute coordinate system

Examples

# Trees relative coordinates
xy <- data.frame(x = runif(200, min = 0, max = 200), y = runif(200, min = 0, max = 200))

# cut the plot in multiple part
coord <- data.frame(X = rep(c(0, 200, 0, 200), 2), Y = rep(c(0, 0, 200, 200), 2))
coord[1:4, ] <- coord[1:4, ] + 5000
coord[5:8, ] <- coord[5:8, ] + 6000
corner <- rep(c(1, 2, 4, 3), 2)
Forestplot <- rep(c("plot1", "plot2"), each = 4)

Outcut <- cutPlot(coord, Forestplot, corner, gridsize = 100, dimX = 200, dimY = 200)

# Assign a plot to 200 trees
Forestplot <- rep(c("plot1", "plot2"), 100)

# attribute trees to subplots
attributeTreeCoord(xy, Forestplot, dim =100,coordAbs = Outcut)
#>        Xproj    Yproj
#> 1   5042.597 5167.164
#> 2   6060.163 6046.990
#> 3   5039.637 5104.927
#> 4   6152.853 6132.822
#> 5   5080.299 5123.718
#> 6   6166.385 6182.008
#> 7   5000.161 5170.831
#> 8   6150.272 6105.736
#> 9   5071.844 5086.099
#> 10  6025.880 6128.059
#> 11  5146.151 5038.452
#> 12  6009.705 6189.596
#> 13  5195.872 5015.274
#> 14  6079.598 6111.326
#> 15  5182.166 5151.876
#> 16  6001.562 6090.016
#> 17  5033.376 5136.537
#> 18  6065.113 6057.002
#> 19  5055.705 5031.395
#> 20  6034.744 6176.223
#> 21  5076.728 5101.582
#> 22  6191.497 6115.279
#> 23  5027.285 5134.815
#> 24  6177.763 6122.528
#> 25  5163.911 5022.853
#> 26  6055.981 6083.637
#> 27  5031.079 5108.498
#> 28  6071.468 6016.729
#> 29  5198.372 5169.127
#> 30  6135.935 6092.063
#> 31  5079.318 5052.425
#> 32  6121.072 6098.785
#> 33  5078.217 5012.866
#> 34  6120.940 6101.870
#> 35  5042.663 5069.453
#> 36  6023.714 6033.662
#> 37  5145.531 5140.896
#> 38  6015.730 6166.185
#> 39  5182.588 5029.450
#> 40  6061.023 6162.407
#> 41  5140.982 5119.908
#> 42  6037.875 6125.753
#> 43  5171.111 5170.440
#> 44  6114.045 6162.203
#> 45  5104.592 5043.914
#> 46  6092.051 6076.735
#> 47  5121.423 5198.133
#> 48  6047.365 6077.898
#> 49  5003.991 5171.840
#> 50  6184.868 6178.208
#> 51  5062.350 5172.480
#> 52  6097.922 6061.566
#> 53  5144.174 5018.645
#> 54  6021.078 6015.199
#> 55  5103.184 5170.442
#> 56  6129.142 6158.438
#> 57  5127.108 5079.867
#> 58  6016.399 6130.246
#> 59  5163.471 5134.845
#> 60  6113.934 6058.062
#> 61  5071.870 5082.301
#> 62  6025.126 6137.283
#> 63  5062.197 5183.604
#> 64  6063.645 6001.143
#> 65  5054.597 5051.102
#> 66  6177.233 6179.190
#> 67  5007.193 5131.404
#> 68  6013.383 6077.516
#> 69  5182.027 5138.815
#> 70  6101.524 6033.815
#> 71  5187.113 5149.276
#> 72  6154.054 6107.287
#> 73  5168.722 5179.711
#> 74  6185.775 6052.774
#> 75  5081.480 5139.505
#> 76  6033.289 6024.625
#> 77  5092.002 5002.384
#> 78  6117.301 6059.087
#> 79  5028.165 5136.146
#> 80  6088.594 6019.825
#> 81  5130.676 5082.521
#> 82  6198.197 6060.823
#> 83  5082.115 5192.784
#> 84  6198.321 6108.773
#> 85  5098.498 5116.309
#> 86  6069.708 6003.252
#> 87  5015.346 5056.952
#> 88  6007.529 6171.709
#> 89  5025.971 5056.499
#> 90  6039.446 6036.644
#> 91  5144.725 5144.591
#> 92  6041.166 6113.544
#> 93  5005.941 5159.583
#> 94  6002.952 6099.946
#> 95  5123.894 5065.796
#> 96  6098.796 6028.402
#> 97  5153.748 5133.316
#> 98  6137.560 6010.346
#> 99  5147.357 5146.256
#> 100 6031.258 6144.171
#> 101 5163.921 5169.082
#> 102 6075.016 6030.541
#> 103 5136.853 5075.269
#> 104 6108.334 6128.223
#> 105 5146.314 5166.499
#> 106 6130.515 6133.437
#> 107 5137.747 5124.483
#> 108 6172.887 6192.964
#> 109 5069.498 5101.788
#> 110 6099.159 6179.291
#> 111 5096.409 5132.842
#> 112 6138.921 6119.462
#> 113 5169.036 5071.366
#> 114 6185.505 6134.959
#> 115 5128.470 5030.210
#> 116 6070.652 6127.990
#> 117 5057.357 5103.426
#> 118 6111.379 6074.946
#> 119 5196.019 5059.330
#> 120 6118.599 6034.646
#> 121 5160.905 5170.492
#> 122 6030.134 6189.724
#> 123 5161.771 5152.214
#> 124 6055.414 6033.266
#> 125 5042.113 5162.980
#> 126 6006.419 6102.169
#> 127 5081.296 5153.483
#> 128 6121.183 6077.354
#> 129 5040.195 5047.031
#> 130 6040.220 6014.230
#> 131 5096.111 5046.640
#> 132 6155.483 6188.857
#> 133 5101.627 5027.942
#> 134 6101.216 6029.309
#> 135 5048.914 5044.347
#> 136 6038.267 6017.367
#> 137 5181.335 5149.913
#> 138 6112.855 6191.973
#> 139 5100.830 5143.897
#> 140 6071.509 6168.272
#> 141 5092.814 5073.209
#> 142 6176.018 6082.779
#> 143 5021.723 5148.101
#> 144 6052.429 6030.878
#> 145 5019.362 5069.154
#> 146 6056.036 6197.300
#> 147 5139.486 5190.900
#> 148 6100.683 6173.705
#> 149 5053.076 5012.470
#> 150 6067.677 6171.974
#> 151 5110.072 5010.977
#> 152 6086.392 6003.358
#> 153 5060.259 5100.482
#> 154 6156.144 6114.301
#> 155 5187.734 5112.549
#> 156 6161.954 6057.779
#> 157 5105.398 5122.518
#> 158 6195.379 6107.317
#> 159 5135.862 5084.805
#> 160 6050.690 6028.812
#> 161 5031.818 5189.348
#> 162 6105.418 6153.092
#> 163 5158.800 5037.154
#> 164 6183.945 6149.961
#> 165 5036.667 5109.229
#> 166 6125.432 6024.403
#> 167 5085.519 5155.342
#> 168 6087.711 6141.131
#> 169 5160.991 5032.361
#> 170 6040.022 6059.929
#> 171 5192.898 5064.138
#> 172 6129.538 6093.801
#> 173 5177.887 5118.964
#> 174 6080.157 6129.837
#> 175 5084.761 5185.267
#> 176 6017.121 6183.824
#> 177 5087.573 5155.729
#> 178 6049.863 6054.028
#> 179 5020.042 5093.857
#> 180 6034.549 6160.278
#> 181 5103.840 5068.915
#> 182 6153.228 6176.534
#> 183 5055.812 5177.053
#> 184 6099.177 6001.814
#> 185 5075.180 5156.422
#> 186 6185.885 6113.986
#> 187 5159.043 5149.946
#> 188 6105.973 6026.922
#> 189 5065.086 5012.648
#> 190 6071.233 6175.800
#> 191 5128.516 5071.570
#> 192 6070.183 6161.706
#> 193 5014.123 5184.646
#> 194 6066.442 6059.976
#> 195 5018.609 5156.912
#> 196 6174.355 6195.292
#> 197 5032.849 5019.022
#> 198 6106.114 6194.284
#> 199 5169.799 5194.262
#> 200 6023.537 6068.630