Attribute trees to GPS coordinates

Attribute trees to GPS coordinates

Usage

attributeTreeCoord(xy, plot, dim, coordAbs)

Arguments

xy

The relative coordinates of the trees within each plot

plot

The label of the plot (same length as the number of rows of xy or length of 1)

dim

The dimension of the plot (either one value if the plot is a square or a vector if a rectangle)

coordAbs

The result of the function cutPlot() or numberCorner()

Value

A data frame with two columns: - Xproj: The X coordinates in the absolute coordinate system - Yproj: The Y coordinates in the absolute coordinate system

Examples

# Trees relative coordinates
xy <- data.frame(x = runif(200, min = 0, max = 200), y = runif(200, min = 0, max = 200))

# cut the plot in multiple part
coord <- data.frame(X = rep(c(0, 200, 0, 200), 2), Y = rep(c(0, 0, 200, 200), 2))
coord[1:4, ] <- coord[1:4, ] + 5000
coord[5:8, ] <- coord[5:8, ] + 6000
corner <- rep(c(1, 2, 4, 3), 2)
Forestplot <- rep(c("plot1", "plot2"), each = 4)

Outcut <- cutPlot(coord, Forestplot, corner, gridsize = 100, dimX = 200, dimY = 200)

# Assign a plot to 200 trees
Forestplot <- rep(c("plot1", "plot2"), 100)

# attribute trees to subplots
attributeTreeCoord(xy, Forestplot, dim =100,coordAbs = Outcut)
#>        Xproj    Yproj
#> 1   5069.518 5167.268
#> 2   6124.289 6158.845
#> 3   5059.177 5175.648
#> 4   6058.182 6181.106
#> 5   5123.784 5154.944
#> 6   6171.179 6033.687
#> 7   5184.044 5177.027
#> 8   6002.586 6096.935
#> 9   5131.305 5062.007
#> 10  6081.247 6054.524
#> 11  5104.871 5110.699
#> 12  6071.291 6155.996
#> 13  5032.619 5135.891
#> 14  6102.037 6079.686
#> 15  5192.752 5182.327
#> 16  6168.173 6123.382
#> 17  5181.256 5133.033
#> 18  6081.395 6111.329
#> 19  5074.307 5083.150
#> 20  6118.298 6123.153
#> 21  5150.199 5148.036
#> 22  6062.617 6113.146
#> 23  5000.413 5151.689
#> 24  6075.882 6023.774
#> 25  5111.587 5197.300
#> 26  6142.506 6066.999
#> 27  5185.416 5061.843
#> 28  6156.189 6155.719
#> 29  5048.574 5035.836
#> 30  6037.497 6192.861
#> 31  5121.651 5075.174
#> 32  6183.472 6011.643
#> 33  5022.945 5099.658
#> 34  6078.103 6152.760
#> 35  5142.283 5001.668
#> 36  6194.406 6195.966
#> 37  5025.885 5131.213
#> 38  6187.751 6095.547
#> 39  5052.832 5164.039
#> 40  6104.734 6162.330
#> 41  5032.577 5065.353
#> 42  6180.631 6056.614
#> 43  5174.077 5184.961
#> 44  6131.468 6151.677
#> 45  5153.761 5018.731
#> 46  6049.047 6039.783
#> 47  5065.649 5031.452
#> 48  6008.995 6037.970
#> 49  5155.635 5179.759
#> 50  6096.610 6094.585
#> 51  5016.769 5060.144
#> 52  6103.360 6141.483
#> 53  5028.458 5068.067
#> 54  6136.729 6134.402
#> 55  5070.391 5171.587
#> 56  6196.790 6138.778
#> 57  5173.414 5173.279
#> 58  6199.008 6157.953
#> 59  5057.909 5030.182
#> 60  6136.685 6086.038
#> 61  5130.715 5087.469
#> 62  6009.425 6190.966
#> 63  5112.817 5158.078
#> 64  6076.210 6195.390
#> 65  5026.483 5094.354
#> 66  6029.137 6052.028
#> 67  5073.231 5193.367
#> 68  6106.111 6062.103
#> 69  5007.869 5179.576
#> 70  6174.910 6134.149
#> 71  5173.837 5111.937
#> 72  6197.701 6133.942
#> 73  5113.530 5180.300
#> 74  6189.310 6074.983
#> 75  5025.414 5088.165
#> 76  6080.064 6074.222
#> 77  5078.242 5010.837
#> 78  6034.688 6064.627
#> 79  5104.899 5089.262
#> 80  6049.920 6117.583
#> 81  5101.193 5146.236
#> 82  6072.561 6060.278
#> 83  5186.146 5123.941
#> 84  6149.943 6086.418
#> 85  5145.980 5059.174
#> 86  6109.322 6043.140
#> 87  5090.499 5099.323
#> 88  6064.339 6125.350
#> 89  5032.067 5198.230
#> 90  6019.953 6092.598
#> 91  5131.153 5182.912
#> 92  6056.215 6168.410
#> 93  5111.227 5032.972
#> 94  6170.655 6040.833
#> 95  5156.925 5180.542
#> 96  6159.733 6054.338
#> 97  5038.551 5053.875
#> 98  6095.339 6073.258
#> 99  5029.245 5092.810
#> 100 6027.443 6072.404
#> 101 5153.637 5062.551
#> 102 6160.829 6148.832
#> 103 5006.691 5161.034
#> 104 6196.392 6154.557
#> 105 5013.885 5147.594
#> 106 6144.018 6132.228
#> 107 5165.819 5194.213
#> 108 6055.125 6190.735
#> 109 5114.515 5177.936
#> 110 6125.116 6091.905
#> 111 5033.761 5106.620
#> 112 6022.053 6117.915
#> 113 5099.659 5089.821
#> 114 6070.725 6122.002
#> 115 5065.467 5047.033
#> 116 6141.946 6185.984
#> 117 5198.475 5095.069
#> 118 6135.829 6149.142
#> 119 5058.507 5047.590
#> 120 6120.372 6183.871
#> 121 5158.312 5074.726
#> 122 6167.397 6011.537
#> 123 5080.470 5056.899
#> 124 6156.301 6076.077
#> 125 5078.514 5117.712
#> 126 6020.142 6083.097
#> 127 5146.313 5117.735
#> 128 6190.575 6073.308
#> 129 5152.603 5164.281
#> 130 6048.159 6067.757
#> 131 5167.195 5146.479
#> 132 6135.084 6058.883
#> 133 5181.537 5034.489
#> 134 6098.370 6048.612
#> 135 5047.504 5035.621
#> 136 6118.701 6053.448
#> 137 5038.825 5099.366
#> 138 6191.589 6135.320
#> 139 5158.683 5034.362
#> 140 6043.251 6188.329
#> 141 5013.100 5113.572
#> 142 6163.634 6182.598
#> 143 5095.919 5119.944
#> 144 6108.000 6076.985
#> 145 5155.917 5014.102
#> 146 6050.101 6000.097
#> 147 5097.416 5133.903
#> 148 6142.650 6069.792
#> 149 5121.524 5060.139
#> 150 6081.173 6101.568
#> 151 5131.192 5010.956
#> 152 6175.470 6182.044
#> 153 5104.802 5174.154
#> 154 6176.728 6045.253
#> 155 5160.524 5142.011
#> 156 6056.381 6089.912
#> 157 5129.994 5067.520
#> 158 6140.107 6078.246
#> 159 5179.098 5189.994
#> 160 6192.961 6076.729
#> 161 5018.541 5183.987
#> 162 6096.769 6076.781
#> 163 5188.574 5064.834
#> 164 6162.867 6062.354
#> 165 5042.358 5088.275
#> 166 6122.043 6143.522
#> 167 5131.981 5176.937
#> 168 6019.475 6130.007
#> 169 5161.644 5162.718
#> 170 6042.420 6180.305
#> 171 5148.633 5075.157
#> 172 6017.599 6075.629
#> 173 5146.307 5148.510
#> 174 6062.391 6024.606
#> 175 5091.146 5177.744
#> 176 6103.594 6116.055
#> 177 5098.475 5085.479
#> 178 6096.216 6154.901
#> 179 5058.351 5186.373
#> 180 6132.609 6023.985
#> 181 5078.592 5044.221
#> 182 6011.699 6104.618
#> 183 5115.277 5163.353
#> 184 6009.002 6050.674
#> 185 5088.270 5030.319
#> 186 6077.419 6154.463
#> 187 5063.268 5114.642
#> 188 6006.019 6007.379
#> 189 5060.885 5130.773
#> 190 6070.265 6199.151
#> 191 5195.775 5064.476
#> 192 6081.992 6071.327
#> 193 5197.893 5166.075
#> 194 6014.859 6017.252
#> 195 5156.709 5015.587
#> 196 6054.323 6193.260
#> 197 5051.486 5052.713
#> 198 6183.150 6106.997
#> 199 5021.939 5103.624
#> 200 6036.756 6174.549