Divides one ore more plots into subplots

This function divides a plot (or several plots) into subplots in the relative coordinates system, and returns the coordinates of subplot corners.

Usage

divide_plot(
  rel_coord,
  proj_coord = NULL,
  grid_size,
  tree_df = NULL,
  tree_coords = NULL,
  corner_plot_ID = NULL,
  tree_plot_ID = NULL,
  grid_tol = 0.1,
  centred_grid = F
)

Arguments

rel_coord

a matrix or data frame containing the relative (local) coordinates of plot corners, with X and Y corresponding to the first and second column respectively

proj_coord

a matrix or data frame containing the projected coordinates of plot corners, with X and Y corresponding to the first and second column respectively, and with the same row order than rel_coord

grid_size

a vector indicating the dimensions of grid cells (resp. X and Y dimensions). If only one value is given, grid cells will be considered as squares.

tree_df

a data frame containing tree relative coordinates and other tree metrics (one row per tree).

tree_coords

a character vector of size 2 containing the column names of the relative coordinates of the trees.

corner_plot_ID

if dealing with multiple plots : a vector indicating plot IDs for corners.

tree_plot_ID

if dealing with multiple plots : a vector indicating tree plot IDs.

grid_tol

a numeric between (0;1) corresponding to the percentage of the plot area allowed to be excluded from the plot division (when grid_size doesn't match exactly plot dimensions).

centred_grid

when grid_size doesn't match exactly plot dimensions, a logical indicating if the subplot grid should be centred on the plot.

Value

If tree_df isn't supplied, returns a data-frame containing as many rows as there are corners corresponding to the subplots, and the following columns :

corner_plot_ID: If dealing with multiple plots : the plot code

subplot_id: The automatically generated subplot code, using the following rule : subplot_X_Y

x_rel and y_rel : the relative X-axis and Y-axis coordinates of subplots corners.

x_proj and y_proj : if proj_coord is supplied, the projected X-axis and Y-axis coordinates of subplots corners

If tree_df is supplied, returns a list containing :

the previous data-frame

the tree_df data-frame with the subplot_id of each tree in the last column

Details

If corner coordinates in the projected coordinate system are supplied (proj_coord), projected coordinates of subplot corners are calculated by a bilinear interpolation in relation with relative coordinates of plot corners. Be aware that this bilinear interpolation only works if the plot in the relative coordinates system is rectangular (ie, has 4 right angles).

Author

Arthur PERE, Arthur BAILLY

Examples

# Rectangular plot and grid cells
rel_coord <- data.frame(x_rel = c(0, 200, 0, 200), y_rel = c(0, 0, 100, 100))
subplots <- divide_plot(rel_coord = rel_coord, grid_size = c(100,50))

# Squared plot and projected coordinates associated
rel_coord <- data.frame(x_rel = c(0, 200, 0, 200), y_rel = c(0, 0, 200, 200))
proj_coord <- data.frame(x_proj = c(210, 383, 110, 283), y_proj = c(210, 310, 383, 483))
subplots <- divide_plot(rel_coord, proj_coord = proj_coord, grid_size = 100)

# Assigning trees to subplots
tree_df <- data.frame(x_tree = runif(50,0,200), y_tree = runif(50,0,200))
subplots <- divide_plot(rel_coord, proj_coord, 100,
                        tree_df = tree_df, tree_coords = c("x_tree","y_tree"))
subplots$sub_corner_coord
#>     subplot_id x_rel y_rel x_proj y_proj
#> 1  subplot_0_0     0     0  210.0  210.0
#> 2  subplot_0_0   100     0  296.5  260.0
#> 3  subplot_0_0   100   100  246.5  346.5
#> 4  subplot_0_0     0   100  160.0  296.5
#> 5  subplot_1_0   100     0  296.5  260.0
#> 6  subplot_1_0   200     0  383.0  310.0
#> 7  subplot_1_0   200   100  333.0  396.5
#> 8  subplot_1_0   100   100  246.5  346.5
#> 9  subplot_0_1     0   100  160.0  296.5
#> 10 subplot_0_1   100   100  246.5  346.5
#> 11 subplot_0_1   100   200  196.5  433.0
#> 12 subplot_0_1     0   200  110.0  383.0
#> 13 subplot_1_1   100   100  246.5  346.5
#> 14 subplot_1_1   200   100  333.0  396.5
#> 15 subplot_1_1   200   200  283.0  483.0
#> 16 subplot_1_1   100   200  196.5  433.0
subplots$tree_df
#>        x_tree     y_tree  subplot_id
#> 1   92.161217  87.826667 subplot_0_0
#> 2   22.394660 121.838824 subplot_0_1
#> 3   53.980142 112.194797 subplot_0_1
#> 4   97.064533 120.880932 subplot_0_1
#> 5  189.082933 184.676845 subplot_1_1
#> 6   69.683542 188.079980 subplot_0_1
#> 7  192.382697 124.244630 subplot_1_1
#> 8   23.342933  97.900301 subplot_0_0
#> 9  128.752198 193.759966 subplot_1_1
#> 10  32.162287  36.250095 subplot_0_0
#> 11 174.424961 125.240717 subplot_1_1
#> 12  72.769263  63.927925 subplot_0_0
#> 13 177.693219  64.367489 subplot_1_0
#> 14 149.289394 107.588101 subplot_1_1
#> 15 111.595751  74.749915 subplot_1_0
#> 16 151.458074  22.586449 subplot_1_0
#> 17 107.059920 172.952964 subplot_1_1
#> 18  30.652118  71.620969 subplot_0_0
#> 19   6.084209 108.057235 subplot_0_1
#> 20  38.593455 173.845049 subplot_0_1
#> 21  49.478390 152.586067 subplot_0_1
#> 22 159.626488  89.802414 subplot_1_0
#> 23  64.763898  19.745450 subplot_0_0
#> 24 199.007480 197.815417 subplot_1_1
#> 25   7.339173  68.712160 subplot_0_0
#> 26   4.344464  36.671449 subplot_0_0
#> 27 169.018924  22.684300 subplot_1_0
#> 28 190.288347 188.320628 subplot_1_1
#> 29 167.602763  70.292833 subplot_1_0
#> 30 172.394021  16.984366 subplot_1_0
#> 31 178.549091 132.253414 subplot_1_1
#> 32 167.544386 198.157239 subplot_1_1
#> 33   5.049280 166.178122 subplot_0_1
#> 34  26.624075  76.144431 subplot_0_0
#> 35  36.438776  84.788411 subplot_0_0
#> 36  13.117103 140.876089 subplot_0_1
#> 37   8.698435  28.334548 subplot_0_0
#> 38 108.515666 113.466192 subplot_1_1
#> 39  50.253620  74.251417 subplot_0_0
#> 40   6.463110 170.803041 subplot_0_1
#> 41  19.030295  84.546999 subplot_0_0
#> 42  16.933631  39.291503 subplot_0_0
#> 43  80.199616  47.233954 subplot_0_0
#> 44  29.454475  23.021458 subplot_0_0
#> 45  53.211666 125.363830 subplot_0_1
#> 46 150.302503  72.323796 subplot_1_0
#> 47 121.148797  36.812470 subplot_1_0
#> 48 133.785693   8.873578 subplot_1_0
#> 49  76.548322 152.111150 subplot_0_1
#> 50 131.955857  13.281403 subplot_1_0

# When grid dimensions don't fit perfectly plot dimensions
subplots <- divide_plot(rel_coord, grid_size = c(45,50))
#> Warning: The x-dimension of the plot is not a multiple of the x-dimension of the grid size
subplots <- divide_plot(rel_coord, grid_size = c(41,41) , grid_tol = 0.4, centred_grid = TRUE)
#> Warning: The x-dimension of the plot is not a multiple of the x-dimension of the grid size
#> Warning: The y-dimension of the plot is not a multiple of the y-dimension of the grid size

# Dealing with multiple plots
rel_coord <- rbind(rel_coord, rel_coord)
proj_coord <- rbind(proj_coord, proj_coord + 200)
tree_df <- rbind(tree_df, data.frame(x_tree = runif(50,0,200), y_tree = runif(50,0,200)))
corner_plot_ID <- rep(c("plot1","plot2"), e=4)
tree_plot_ID <- rep(c("plot1","plot2"), e=50)
subplots <- divide_plot(rel_coord, proj_coord, 100,
                        tree_df, c("x_tree","y_tree"),
                        corner_plot_ID, tree_plot_ID)