Divides one ore more plots into subplots

This function divides a plot (or several plots) into subplots in the relative coordinates system, and returns the coordinates of subplot corners.

Usage

divide_plot(
  rel_coord,
  proj_coord = NULL,
  grid_size,
  tree_df = NULL,
  tree_coords = NULL,
  corner_plot_ID = NULL,
  tree_plot_ID = NULL,
  grid_tol = 0.1,
  centred_grid = F
)

Arguments

rel_coord

a data frame containing the relative (local) coordinates of plot corners, with X and Y corresponding to the first and second column respectively

proj_coord

a data frame containing the projected coordinates of plot corners, with X and Y corresponding to the first and second column respectively, and with the same row order than rel_coord

grid_size

a vector indicating the dimensions of grid cells (resp. X and Y dimensions). If only one value is given, grid cells will be considered as squares.

tree_df

a data frame containing tree relative coordinates and other tree metrics (one row per tree).

tree_coords

a character vector of size 2 containing the column names of the relative coordinates of the trees.

corner_plot_ID

if dealing with multiple plots : a vector indicating plot IDs for corners.

tree_plot_ID

if dealing with multiple plots : a vector indicating tree plot IDs.

grid_tol

a numeric between (0;1) corresponding to the percentage of the plot area allowed to be excluded from the plot division (when grid_size doesn't match exactly plot dimensions).

centred_grid

when grid_size doesn't match exactly plot dimensions, a logical indicating if the subplot grid should be centred on the plot.

Value

If tree_df isn't supplied, returns a data-frame containing as many rows as there are corners corresponding to the subplots, and the following columns :

corner_plot_ID: If dealing with multiple plots : the plot code

subplot_id: The automatically generated subplot code, using the following rule : subplot_X_Y

x_rel and y_rel : the relative X-axis and Y-axis coordinates of subplots corners.

x_proj and y_proj : if proj_coord is supplied, the projected X-axis and Y-axis coordinates of subplots corners

If tree_df is supplied, returns a list containing :

the previous data-frame

the tree_df data-frame with the subplot_id of each tree in the last column

Details

If corner coordinates in the projected coordinate system are supplied (proj_coord), projected coordinates of subplot corners are calculated by a bilinear interpolation in relation with relative coordinates of plot corners. Be aware that this bilinear interpolation only works if the plot in the relative coordinates system is rectangular (ie, has 4 right angles).

Author

Arthur PERE, Arthur BAILLY

Examples

# Rectangular plot and grid cells
rel_coord <- data.frame(x_rel = c(0, 200, 0, 200), y_rel = c(0, 0, 100, 100))
subplots <- divide_plot(rel_coord = rel_coord, grid_size = c(100,50))

# Squared plot and projected coordinates associated
rel_coord <- data.frame(x_rel = c(0, 200, 0, 200), y_rel = c(0, 0, 200, 200))
proj_coord <- data.frame(x_proj = c(110, 190, 60, 145), y_proj = c(110, 160, 196, 245))
subplots <- divide_plot(rel_coord, proj_coord = proj_coord, grid_size = 100)

# Assigning trees to subplots
tree_df <- data.frame(x_tree = runif(50,0,200), y_tree = runif(50,0,200))
subplots <- divide_plot(rel_coord, proj_coord, 100, tree_df = tree_df, tree_coords = c("x_tree","y_tree"))
subplots$sub_corner_coord
#>     subplot_id x_rel y_rel x_proj y_proj
#> 1  subplot_0_0     0     0 110.00 110.00
#> 2  subplot_0_0   100     0 150.00 135.00
#> 3  subplot_0_0   100   100 126.25 177.75
#> 4  subplot_0_0     0   100  85.00 153.00
#> 5  subplot_1_0   100     0 150.00 135.00
#> 6  subplot_1_0   200     0 190.00 160.00
#> 7  subplot_1_0   200   100 167.50 202.50
#> 8  subplot_1_0   100   100 126.25 177.75
#> 9  subplot_0_1     0   100  85.00 153.00
#> 10 subplot_0_1   100   100 126.25 177.75
#> 11 subplot_0_1   100   200 102.50 220.50
#> 12 subplot_0_1     0   200  60.00 196.00
#> 13 subplot_1_1   100   100 126.25 177.75
#> 14 subplot_1_1   200   100 167.50 202.50
#> 15 subplot_1_1   200   200 145.00 245.00
#> 16 subplot_1_1   100   200 102.50 220.50
subplots$tree_df
#>         x_tree     y_tree  subplot_id
#> 1    1.1841259 142.885723 subplot_0_1
#> 2    6.3336429 102.800312 subplot_0_1
#> 3   35.6989492  50.388549 subplot_0_0
#> 4  177.4095955 148.722249 subplot_1_1
#> 5   67.3851625   2.058847 subplot_0_0
#> 6   65.0711877  93.492055 subplot_0_0
#> 7   53.1773906 113.717417 subplot_0_1
#> 8   32.0249834 134.544908 subplot_0_1
#> 9  136.9999903 128.393741 subplot_1_1
#> 10 162.3749448 111.926955 subplot_1_1
#> 11  89.8975748  97.302036 subplot_0_0
#> 12  50.6842533  28.844832 subplot_0_0
#> 13 111.4264535  97.840742 subplot_1_0
#> 14  31.6954310 194.278475 subplot_0_1
#> 15 130.4117911 143.157408 subplot_1_1
#> 16  52.5944006  40.847633 subplot_0_0
#> 17   2.2896021  86.096922 subplot_0_0
#> 18 190.0319621  48.245051 subplot_1_0
#> 19   0.1593006  65.175159 subplot_0_0
#> 20  93.4226763  32.606683 subplot_0_0
#> 21  98.5297266  30.246337 subplot_0_0
#> 22  38.0014177  70.536619 subplot_0_0
#> 23  15.5833755 146.645141 subplot_0_1
#> 24 104.2204365  31.404474 subplot_1_0
#> 25  19.8484642  11.598530 subplot_0_0
#> 26  82.2074093  47.316358 subplot_0_0
#> 27 192.1545198 173.030675 subplot_1_1
#> 28  79.7505928 125.914499 subplot_0_1
#> 29  26.9077156 151.241623 subplot_0_1
#> 30  64.9379437  73.753701 subplot_0_0
#> 31 131.2872568 185.390672 subplot_1_1
#> 32 190.6540796 126.707831 subplot_1_1
#> 33 185.6232701 198.015831 subplot_1_1
#> 34  55.6408145 195.775086 subplot_0_1
#> 35  81.0492968 194.338626 subplot_0_1
#> 36  62.6439699 199.238738 subplot_0_1
#> 37  71.1396301  76.407751 subplot_0_0
#> 38  17.7798766 104.653136 subplot_0_1
#> 39 115.1623120 131.014527 subplot_1_1
#> 40 130.1040413 156.012245 subplot_1_1
#> 41 155.1882199 103.188538 subplot_1_1
#> 42  42.9845494  75.265072 subplot_0_0
#> 43 147.5220324  74.788681 subplot_1_0
#> 44  22.8338909 105.407252 subplot_0_1
#> 45  38.2252812 105.495410 subplot_0_1
#> 46  25.3960830  68.445088 subplot_0_0
#> 47   5.4649187 105.692037 subplot_0_1
#> 48   6.2730053 120.132518 subplot_0_1
#> 49  59.1555755   2.253360 subplot_0_0
#> 50  95.5083790 139.634523 subplot_0_1

# When grid dimensions don't fit perfectly plot dimensions
subplots <- divide_plot(rel_coord, grid_size = c(45,50))
#> Warning: The x-dimension of the plot is not a multiple of the x-dimension of the grid size
subplots <- divide_plot(rel_coord, grid_size = c(41,41) , grid_tol = 0.4, centred_grid = TRUE)
#> Warning: The x-dimension of the plot is not a multiple of the x-dimension of the grid size
#> Warning: The y-dimension of the plot is not a multiple of the y-dimension of the grid size

# Dealing with multiple plots
rel_coord <- rbind(rel_coord, rel_coord)
proj_coord <- rbind(proj_coord, proj_coord + 200)
tree_df <- rbind(tree_df, data.frame(x_tree = runif(50,0,200), y_tree = runif(50,0,200)))
corner_plot_ID <- rep(c("plot1","plot2"), e=4)
tree_plot_ID <- rep(c("plot1","plot2"), e=50)
subplots <- divide_plot(rel_coord, proj_coord, 100, tree_df, c("x_tree","y_tree"), corner_plot_ID, tree_plot_ID)