Summarise and display tree information by subplot

After applying the divide_plot() function, this function summarises with any defined function the desired tree metric by sub-plot and displays the plot representation.

Usage

subplot_summary(subplots, value = NULL, draw_plot = TRUE, fun = sum, ...)

Arguments

subplots

output of the divide_plot() function

value

a character indicating the column in subplots$tree_df to display

draw_plot

a logical indicating whether the plot design should be displayed and returned

fun

the function to be applied

...

optional arguments to fun

Value

a list containg the following elements :

tree_summary : a summary of the metric per subplot

polygon : an sf object : simple feature collection of the subplot's polygon

plot_design : (if draw_plot=T) a ggplot object (or a list of ggplot objects) that can easily be modified

Author

Arthur Bailly

Examples

rel_coord <- data.frame(x_rel = c(0, 200, 0, 200), y_rel = c(0, 0, 200, 200))
proj_coord <- data.frame(x_proj = c(210, 383, 110, 283), y_proj = c(210, 310, 383, 483))
tree_df <- data.frame(x_tree = runif(50,0,200), y_tree = runif(50,0,200), metric = rnorm(50,10,5))
subplots <- divide_plot(rel_coord, proj_coord = proj_coord, grid_size = 50,
                        tree_df = tree_df, tree_coords = c("x_tree","y_tree"))
# Sum summary (by default)
subplot_summary(subplots , value = "metric", draw_plot = FALSE)
#> $tree_summary
#>     subplot_id metric_summary metric_summary_per_ha
#> 1  subplot_0_1       40.36589             161.75067
#> 2  subplot_0_2       44.11341             176.76742
#> 3  subplot_0_3       27.04323             108.36525
#> 4  subplot_1_0       28.82499             115.50498
#> 5  subplot_1_1       10.45078              41.87745
#> 6  subplot_1_2       49.82367             199.64904
#> 7  subplot_1_3       23.71833              95.04201
#> 8  subplot_2_0       73.66027             295.16501
#> 9  subplot_2_1       10.42498              41.77409
#> 10 subplot_2_2       19.25030              77.13813
#> 11 subplot_2_3       20.60999              82.58657
#> 12 subplot_3_0       57.14601             228.99049
#> 13 subplot_3_1       52.45570             210.19588
#> 14 subplot_3_2       10.04750              40.26147
#> 15 subplot_3_3       25.18808             100.93146
#> 
#> $polygon
#> Simple feature collection with 16 features and 3 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 110 ymin: 210 xmax: 383 ymax: 483
#> CRS:           NA
#> First 10 features:
#>              subplot_id metric_summary metric_summary_per_ha
#> subplot_0_0 subplot_0_0        0.00000               0.00000
#> subplot_0_1 subplot_0_1       40.36589             161.75067
#> subplot_0_2 subplot_0_2       44.11341             176.76742
#> subplot_0_3 subplot_0_3       27.04323             108.36525
#> subplot_1_0 subplot_1_0       28.82499             115.50498
#> subplot_1_1 subplot_1_1       10.45078              41.87745
#> subplot_1_2 subplot_1_2       49.82367             199.64904
#> subplot_1_3 subplot_1_3       23.71833              95.04201
#> subplot_2_0 subplot_2_0       73.66027             295.16501
#> subplot_2_1 subplot_2_1       10.42498              41.77409
#>                         sf_subplot_polygon
#> subplot_0_0 POLYGON ((210 210, 253.25 2...
#> subplot_0_1 POLYGON ((185 253.25, 228.2...
#> subplot_0_2 POLYGON ((160 296.5, 203.25...
#> subplot_0_3 POLYGON ((135 339.75, 178.2...
#> subplot_1_0 POLYGON ((253.25 235, 296.5...
#> subplot_1_1 POLYGON ((228.25 278.25, 27...
#> subplot_1_2 POLYGON ((203.25 321.5, 246...
#> subplot_1_3 POLYGON ((178.25 364.75, 22...
#> subplot_2_0 POLYGON ((296.5 260, 339.75...
#> subplot_2_1 POLYGON ((271.5 303.25, 314...
#> 
# 9th quantile summary (for example)
subplot_summary(subplots , value = "metric", draw_plot = FALSE,
                fun = quantile, probs=0.9)
#> $tree_summary
#>     subplot_id metric_summary metric_summary_per_ha
#> 1  subplot_0_1      14.509661              58.14184
#> 2  subplot_0_2      16.260469              65.15753
#> 3  subplot_0_3      10.494205              42.05146
#> 4  subplot_1_0      12.321004              49.37165
#> 5  subplot_1_1      10.450781              41.87745
#> 6  subplot_1_2      15.771173              63.19686
#> 7  subplot_1_3      13.723469              54.99148
#> 8  subplot_2_0      17.909439              71.76514
#> 9  subplot_2_1       9.126446              36.57070
#> 10 subplot_2_2      12.337602              49.43816
#> 11 subplot_2_3      20.609994              82.58657
#> 12 subplot_3_0      14.400939              57.70618
#> 13 subplot_3_1      17.869030              71.60322
#> 14 subplot_3_2      10.047502              40.26147
#> 15 subplot_3_3       8.253866              33.07417
#> 
#> $polygon
#> Simple feature collection with 16 features and 3 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 110 ymin: 210 xmax: 383 ymax: 483
#> CRS:           NA
#> First 10 features:
#>              subplot_id metric_summary metric_summary_per_ha
#> subplot_0_0 subplot_0_0       0.000000               0.00000
#> subplot_0_1 subplot_0_1      14.509661              58.14184
#> subplot_0_2 subplot_0_2      16.260469              65.15753
#> subplot_0_3 subplot_0_3      10.494205              42.05146
#> subplot_1_0 subplot_1_0      12.321004              49.37165
#> subplot_1_1 subplot_1_1      10.450781              41.87745
#> subplot_1_2 subplot_1_2      15.771173              63.19686
#> subplot_1_3 subplot_1_3      13.723469              54.99148
#> subplot_2_0 subplot_2_0      17.909439              71.76514
#> subplot_2_1 subplot_2_1       9.126446              36.57070
#>                         sf_subplot_polygon
#> subplot_0_0 POLYGON ((210 210, 253.25 2...
#> subplot_0_1 POLYGON ((185 253.25, 228.2...
#> subplot_0_2 POLYGON ((160 296.5, 203.25...
#> subplot_0_3 POLYGON ((135 339.75, 178.2...
#> subplot_1_0 POLYGON ((253.25 235, 296.5...
#> subplot_1_1 POLYGON ((228.25 278.25, 27...
#> subplot_1_2 POLYGON ((203.25 321.5, 246...
#> subplot_1_3 POLYGON ((178.25 364.75, 22...
#> subplot_2_0 POLYGON ((296.5 260, 339.75...
#> subplot_2_1 POLYGON ((271.5 303.25, 314...
#>